mingw-w64-luajit-luarocks: add
This package is based on "mingw-w64-lua-luarocks".

[LuaJIT](https://luajit.org/luajit.html) is a JIT(Just-in-Time
Compiler) for the Lua 5.1 compatible.
This makes Lua program faster than the original Lua interpreter.

Therefore, it is useful for many users that
we can use LuaRocks and Lua modules that are installed from LuaRocks
on LuaJIT.

Since LuaJIT conflicts with Lua 5.1, we need a new package.

Lua 5.1 is needed to build "mingw-w64-lua51-luarocks". And also
LuaJIT is needed to build "mingw-w64-luajit-luarocks".

However, Lua 5.1 and LuaJIT should be conflict because LuaJIT PKGBUILD
marks conflict for Lua 5.1.

So, "mingw-w64-lua51-luarocks" and "mingw-w64-luajit-luarocks" won't
be able to built in the same environment.

The difference with "minge-w64-lua-luarocks" is as below.

  * pkgver:

    * Change to 3.3.1 (This is a latest version for now.) from 2.2.4.

  * pkgname:

    * Change to "mingw-w64-x86_x64-luajit-3.3.1" from "mingw-w64-x86_x64-lua51-2.2.4".

  * depends:

    * Change to "mingw-w64-x86_64-luajit" from "mingw-w64-x86_64-lua51".

  * conflicts:

    * Add a "mingw-w64-x86_64-lua51-luarocks".

  * configure options:

    * lua include path:

      * Change to
        "--with-lua-include=${MINGW_PREFIX}/include/luajit-2.0" from
        "--with-lua-include=${MINGW_PREFIX}/include/lua5.1".

    * lua suffix:

      * Change to "--lua-suffix=jit.exe" from "--lua-suffix=5.1.exe".

    * lua version:

      * Remove a "--lua-version=5.1" option.
        Because if we specify 5.1 into this option, the luarocks can't found the luajit.

  * post install:

    * the path of lua directory change to path for Windows in luarocks
      binary(bin/luarocks, bin/luarocks-5.1, bin/luarocks-admin, and bin/luarocks-admin-5.1).

  * mingw32-msys2.patch:

    * defaults.variables.LUALIB:

      * Change to "libluajit-"..cfg.lua_version..".dll.a" from "liblua"..cfg.lua_version..".dll.a".